cleaned up code and output methods, enhanced docu
