+++ cloudcost.gemspec
@@ -10,8 +10,7 @@ Gem::Specification.new do |s|
 
   s.required_rubygems_version = Gem::Requirement.new(">= 0") if s.respond_to? :required_rubygems_version=
   s.authors = ["Nik Wolfgramm"]
-  s.description = "Calculate cloudscale.ch server costs from your actual deployment"
-  s.email = "wolfgramm@puzzle.ch"
+  s.description = "Calculate cloudscale.ch server costs from the current deployment"
   s.files = `git ls-files`.split($INPUT_RECORD_SEPARATOR)
   s.require_paths = ["lib"]
   s.required_ruby_version = ">= 2.7"
@@ -20,7 +19,7 @@ Gem::Specification.new do |s|
   s.test_files = s.files.grep(%r{^(test|spec|features)/})
   s.license = "MIT"
 
-  s.add_dependency("excon", "~> 0.82.0")
+  s.add_dependency("excon", "~> 0.85.0")
   s.add_dependency("parseconfig", "~> 1.1.0")
   s.add_dependency("terminal-table", "~> 3.0.1")
   s.add_dependency("thor", "~> 1.1.0")
