+++ lib/cloudcost.rb
@@ -7,5 +7,8 @@ require "cloudcost/version"
 require "cloudcost/api_token"
 require "cloudcost/api_connection"
 require "cloudcost/server"
+require "cloudcost/tabular_output"
+require "cloudcost/influxdb_output"
+require "cloudcost/csv_output"
 require "cloudcost/server_list"
 require "cloudcost/cli"
